Separate TodosHeadersAdapter into its own file.
